Ejemplo de servicios REST para eventos. Falta llave foranea a sensores
