+++ event_manager/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from .models import Event
+
+class EventSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Event
+        fields = ('id', 'timestamp')
