+++ event_manager/views.py
@@ -1,0 +1,13 @@
+from django.shortcuts import render
+
+# Create your views here.
+from rest_framework import viewsets
+from .models import Event
+from .serializers import EventSerializer
+
+class EventViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows users to be viewed or edited.
+    """
+    queryset = Event.objects.all()
+    serializer_class = EventSerializer
